% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_streamorder-level.R
\name{add_streamorder}
\alias{add_streamorder}
\alias{add_streamorder.data.frame}
\alias{add_streamorder.hy}
\alias{add_streamorder.hy_node}
\alias{add_streamorder.hy_flownetwork}
\alias{add_streamorder.hy_topo}
\title{Add Streamorder}
\usage{
add_streamorder(x, status = TRUE)

\method{add_streamorder}{data.frame}(x, status = TRUE)

\method{add_streamorder}{hy}(x, status = TRUE)

\method{add_streamorder}{hy_node}(x, status = TRUE)

\method{add_streamorder}{hy_flownetwork}(x, status = TRUE)

\method{add_streamorder}{hy_topo}(x, status = TRUE)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{status}{boolean if status updates should be printed.}
}
\value{
data.frame containing added \code{stream_order} and \code{stream_calculator} attribute.
}
\description{
Adds a strahler stream order.

Algorithm: If more than one upstream flowline has an order equal to the
maximum upstream order then the downstream flowline is assigned the maximum
upstream order plus one. Otherwise it is assigned the maximum upstream order.

To match the NHDPlus algorithm, non-dendritic network connectivity must
be included. All secondary paths will have the \code{stream_order} of upstream
primary paths and a \code{stream_calculator} value of 0. Secondary paths have
no affect on the order of downstream paths.
}
\details{
Required attributes: \code{id} and \code{toid} or \code{fromnode}, \code{tonode}, and \code{divergence}
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

x <- dplyr::select(x, COMID, FromNode, ToNode, Divergence)

x <- add_streamorder(x)

plot(sf::st_geometry(x), lwd = x$stream_order, col = "blue")
plot(sf::st_geometry(x), lwd = x$stream_calculator, col = "blue")

}
