% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_network.R
\name{add_topo_sort}
\alias{add_topo_sort}
\alias{add_topo_sort.data.frame}
\alias{add_topo_sort.hy}
\alias{add_topo_sort.hy_node}
\alias{add_topo_sort.hy_flownetwork}
\alias{add_topo_sort.hy_topo}
\title{Add topo_sort}
\usage{
add_topo_sort(x, outlets = NULL)

\method{add_topo_sort}{data.frame}(x, outlets = NULL)

\method{add_topo_sort}{hy}(x, outlets = NULL)

\method{add_topo_sort}{hy_node}(x, outlets = NULL)

\method{add_topo_sort}{hy_flownetwork}(x, outlets = NULL)

\method{add_topo_sort}{hy_topo}(x, outlets = NULL)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{outlets}{same as id in x. if specified, only the network
emanating from these outlets will be considered and returned.}
}
\value{
data.frame containing a topo_sort attribute.
}
\description{
calls \link{sort_network} without support for splitting the network
and adds a \code{nrow:1} topo_sort attribute.
}
\details{
Required attributes: \code{id}, \code{toid}
}
