% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_names.R
\name{align_names}
\alias{align_names}
\title{Align names to hydroloom convention}
\usage{
align_names(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
data.frame renamed to match hydroloom as possible.
}
\description{
this function aligns the attribute names
with those used in hydroloom.
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

names(x)

x <- align_names(x)

names(x)

}
