% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfs.R
\name{all_paths_dfs}
\alias{all_paths_dfs}
\title{all paths depth first}
\usage{
all_paths_dfs(x, starts, direction = "down")
}
\arguments{
\item{x}{hy data.frame containing network topology}

\item{starts}{vector with ids from x to start at.}

\item{direction}{character only "down" supported so far.}
}
\description{
given a starting node, return all reachable paths
}
\examples{

x <- hy(sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom")))

x <- add_toids(x, return_dendritic = FALSE)

all_paths_dfs(x, 8893402)

}
