% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_geometry}
\alias{drop_geometry}
\title{drop_geometry}
\usage{
drop_geometry(x)
}
\arguments{
\item{x}{data.frame that may contain a geometry column}
}
\value{
data.frame without geometry column
}
\description{
drops geometry if present, does nothing otherwise.
}
\examples{

(g <- sf::st_sf(a=3, geo = sf::st_sfc(sf::st_point(1:2))))
drop_geometry(g)

}
