% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_articulation.R
\name{get_articulation_flowlines}
\alias{get_articulation_flowlines}
\alias{get_articulation_flowlines.data.frame}
\alias{get_articulation_flowlines.hy}
\title{Get Articulation Flowlines}
\usage{
get_articulation_flowlines(x)

\method{get_articulation_flowlines}{data.frame}(x)

\method{get_articulation_flowlines}{hy}(x)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}
}
\value{
vector of flowline ids that are articulation points in the network
}
\description{
Identifies articulation flowlines (cut vertices) in the network.
Articulation flowlines are those whose removal would disconnect the network.
}
\examples{

x <- data.frame(
  id = c(1, 2, 3, 4, 5, 6, 7, 8, 9),
  toid = c(2, 3, 4, 5, 0, 7, 8, 9, 4)
)

# 1 -> 2 -> 3 -> 4 -> 5
#               ^
#               |
# 6 -> 7 -> 8 -> 9

get_articulation_flowlines(x)

}
