% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_containment.R
\name{get_containing_domain}
\alias{get_containing_domain}
\title{Look up a domain's containing domain}
\usage{
get_containing_domain(decomposition, domain_id)
}
\arguments{
\item{decomposition}{object of class \code{domain_decomposition}.}

\item{domain_id}{scalar or vector of domain ids (keys of
\code{decomposition$domains}).}
}
\value{
character vector of containing domain ids, same length as
\code{domain_id}. \code{NA_character_} for domains that have no container.
}
\description{
Returns the \code{containing_domain_id} slot for one or more domains in a
decomposition. Containment is declared via \code{\link[=set_containment]{set_containment()}};
domains without a container return \code{NA_character_}.
}
\examples{
g <- sf::read_sf(system.file("extdata/walker.gpkg", package = "hydroloom"))

h <- hy(g) |>
  add_toids() |>
  add_levelpaths(name_attribute = "GNIS_ID",
    weight_attribute = "arbolate_sum")

d <- decompose_network(h)

get_containing_domain(d, names(d$domains)[1])

}
\seealso{
\code{\link[=set_containment]{set_containment()}}, \code{\link[=get_domain_graph]{get_domain_graph()}},
\link{domain_decomposition}.
}
