% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_network.R
\name{get_overrides}
\alias{get_overrides}
\title{Get the overrides table from a decomposition}
\usage{
get_overrides(decomposition)
}
\arguments{
\item{decomposition}{object of class \code{domain_decomposition}.}
}
\value{
data.frame or \code{NULL}.
}
\description{
Returns \code{decomposition$overrides} -- the non-dendritic inter-domain
transfer table passed through from \code{\link[=decompose_network]{decompose_network()}}. \code{NULL}
when no overrides were supplied.
}
\examples{
g <- sf::read_sf(system.file("extdata/walker.gpkg", package = "hydroloom"))

h <- hy(g) |>
  add_toids() |>
  add_levelpaths(name_attribute = "GNIS_ID",
    weight_attribute = "arbolate_sum")

d <- decompose_network(h)

get_overrides(d)

}
\seealso{
\link{domain_decomposition}, \code{\link[=decompose_network]{decompose_network()}}.
}
