% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_hydroloom.R
\name{hy_reverse}
\alias{hy_reverse}
\title{reverse hy to original names}
\usage{
hy_reverse(x)
}
\arguments{
\item{x}{hy data.frame to be returned to original state}
}
\description{
renames hy object to original names and removes hy object
attributes.
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))
x <- hy(x)

hy_reverse(x)

}
