% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_names.R
\name{hydroloom_names}
\alias{hydroloom_names}
\title{Get or Set Hydroloom Names}
\usage{
hydroloom_names(x = NULL, clear = FALSE)
}
\arguments{
\item{x}{named character vector of additional names to add to the
hydroloom environment. If not specified, no names will be added and
the current value stored in the hydroloom environment will be returned.}

\item{clear}{logical if TRUE, all names will be removed and replaced with
x.}
}
\value{
named character vector containing hydroloom names with registered attribute name mappings in \code{names}.
}
\description{
Retrieve hydroloom name mapping from hydroloom
environment. Hydroloom uses a specific set of attribute names within
the package and includes mappings from names used in some data sources.
This function will return those names and can be used to set additional
name mappings.

NOTE: these values will reset when R is restarted. Add desired settings
to a project or user .Rprofile to make long term additions.
}
\examples{

hydroloom_names()

}
