% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy.R
\name{is.hy}
\alias{is.hy}
\title{Is Valid \code{hy} Class?}
\usage{
is.hy(x, silent = FALSE)
}
\arguments{
\item{x}{object to test}

\item{silent}{logical should messages be emitted?}
}
\value{
logical TRUE if valid
}
\description{
test if object is a valid according to the hy s3 class
}
