% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_network.R
\name{is_leaf_domain}
\alias{is_leaf_domain}
\title{Test whether a domain is a leaf}
\usage{
is_leaf_domain(decomposition, domain_id)
}
\arguments{
\item{decomposition}{object of class \code{domain_decomposition}.}

\item{domain_id}{character(1). Domain id to test.}
}
\value{
logical(1).
}
\description{
A domain is a \emph{leaf} when no upstream domain feeds into it — its
\code{inlet_nexus_ids} slot is empty.
}
\examples{
g <- sf::read_sf(system.file("extdata/walker.gpkg", package = "hydroloom"))

h <- hy(g) |>
  add_toids() |>
  add_levelpaths(name_attribute = "GNIS_ID",
    weight_attribute = "arbolate_sum")

d <- decompose_network(h)

is_leaf_domain(d, names(d$domains)[[1]])

}
\seealso{
\code{\link[=is_stem_domain]{is_stem_domain()}}, \code{\link[=is_root_domain]{is_root_domain()}}, \code{\link[=hy_domain]{hy_domain()}}.
}
