% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_geometry}
\alias{rename_geometry}
\title{Rename Geometry}
\usage{
rename_geometry(g, name)
}
\arguments{
\item{g}{sf data.table}

\item{name}{character name to be used for geometry}
}
\value{
sf data.frame with geometry column renamed according to name parameter
}
\description{
correctly renames the geometry column
of a sf object.
}
\examples{

(g <- sf::st_sf(a = 3, geo = sf::st_sfc(sf::st_point(1:2))))
rename_geometry(g, "geometry")

}
