% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_divergence.R
\name{add_return_divergence}
\alias{add_return_divergence}
\alias{add_return_divergence.data.frame}
\title{add return divergence}
\usage{
add_return_divergence(x, status = TRUE)

\method{add_return_divergence}{data.frame}(x, status = TRUE)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{status}{boolean if status updates should be printed.}
}
\value{
data.frame containing \code{return_divergence} attribute
}
\description{
Adds a return divergence attribute
Algorithm: TODO
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

x <- hy(x)

x <- add_return_divergence(x)

}
