% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_order.R
\name{add_streamorder}
\alias{add_streamorder}
\alias{add_streamorder.data.frame}
\title{add Streamorder}
\usage{
add_streamorder(x, status = TRUE)

\method{add_streamorder}{data.frame}(x, status = TRUE)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{status}{boolean if status updates should be printed.}
}
\value{
numeric stream order in same order as input
}
\description{
Adds a strahler stream order.
Algorithm: If more than one upstream flowpath has an order equal to the
maximum upstream order then the downstream flowpath is assigned the maximum
upstream order plus one. Otherwise it is assigned the max upstream order.
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

x <- add_toids(x)

x <- add_streamorder(x)

plot(sf::st_geometry(x), lwd = x$order, col = "blue")

}
