% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{add_toids}
\alias{add_toids}
\title{add toids}
\usage{
add_toids(x, return_dendritic = TRUE)
}
\arguments{
\item{x}{hy object without a toid attribute}

\item{return_dendritic}{logical remove non dendritic paths if TRUE. Requires
a "divergence" flag where 1 is main and 2 is secondary.}
}
\value{
hy object with toid attribute
}
\description{
Given an hy object with fromnode and tonode attributes,
will return a toid attribute that is the result of joining
tonode and fromnode attributes.
}
\examples{
x <- hy(sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom")))

x <- add_toids(x)

}
