% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hy_graph.R
\name{check_hy_graph}
\alias{check_hy_graph}
\title{Check hy graph}
\usage{
check_hy_graph(x, loop_check = FALSE)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{loop_check}{logical if TRUE, a special mode of
\link{navigate_network_dfs} is used to walk the entire network from
top to bottom searching for loops.}
}
\value{
if no localized loops are found, returns TRUE. If localized
loops are found, problem rows with a row number added.
}
\description{
check that a id toid graph doesn't contain localized loops.
}
\examples{
# notice that row 4 (id = 4, toid = 9) and row 8 (id = 9, toid = 4) is a loop.
test_data <- data.frame(id = c(1, 2, 3, 4, 6, 7, 8, 9),
                      toid = c(2, 3, 4, 9, 7, 8, 9, 4))
check_hy_graph(test_data)

}
