% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_ids.R
\name{check_hy_graph}
\alias{check_hy_graph}
\title{Check hy graph}
\usage{
check_hy_graph(x)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}
}
\value{
if no localized loops are found, returns TRUE. If localized
loops are found, problem rows with a row number added.
}
\description{
check that a id toid graph doesn't contain localized loops.
}
