% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_ids.R
\name{format_index_ids}
\alias{format_index_ids}
\title{format index ids}
\usage{
format_index_ids(g, return_list = FALSE)
}
\arguments{
\item{g}{data.frame with \code{inid} and \code{toindid} as returned by \link{make_index_ids}
with \code{long_form}=TRUE.}

\item{complete}{logical should the a data.frame with a list column be included
in the return?}
}
\value{
list containing an adjacency matrix and a lengths vector indicating
the number of connections from each node. If \code{complete} is \code{TRUE} return
will also include a data.frame with an \code{indid} column and a \code{toindid} list
column.
}
\description{
format index ids
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

y <- add_toids(x) |>
  make_index_ids(long_form = TRUE) |>
  format_index_ids()

}
