% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_network.R
\name{get_domain_for_catchment}
\alias{get_domain_for_catchment}
\title{Look up the domain containing a catchment}
\usage{
get_domain_for_catchment(decomposition, catchment_id)
}
\arguments{
\item{decomposition}{object of class \code{domain_decomposition}.}

\item{catchment_id}{scalar or vector of catchment ids.}
}
\value{
character vector of domain ids, same length as
\code{catchment_id}.
}
\description{
Returns the \code{domain_id} of the domain that owns a given catchment in
a decomposition. Accepts a scalar or vector of catchment ids.
}
\examples{
g <- sf::read_sf(system.file("extdata/walker.gpkg", package = "hydroloom"))

h <- hy(g) |>
  add_toids() |>
  add_levelpaths(name_attribute = "GNIS_ID",
    weight_attribute = "arbolate_sum")

d <- decompose_network(h)

get_domain_for_catchment(d, h$id[1])

}
\seealso{
\code{\link[=decompose_network]{decompose_network()}}.
}
