% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroloom.R
\name{hy}
\alias{hy}
\title{create an hy fabric object}
\usage{
hy(x)
}
\arguments{
\item{x}{data.frame with compatible attribute names from nhdplus}
}
\description{
converts a compatible dataset into a fabric s3 class
}
\examples{
library(nhdplusTools)
source(system.file("extdata/nhdplushr_data.R", package = "nhdplusTools"))
nhd <- hr_data$NHDFlowline
hy(nhd)
}
