% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydroloom.R
\name{hy}
\alias{hy}
\title{create an hy fabric object}
\usage{
hy(x, clean = FALSE)
}
\arguments{
\item{x}{data.frame with compatible attribute names from nhdplus}
}
\value{
hy object with attributes compatible with the hydroloom package.
}
\description{
converts a compatible dataset into a fabric s3 class
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))
x <- hy(x)

}
