% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy.R
\name{hy}
\alias{hy}
\title{Create a hy Fabric S3 Object}
\usage{
hy(x, clean = FALSE)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{clean}{logical if TRUE, geometry and non-hydroloom compatible attributes
will be removed.}
}
\value{
hy object with attributes compatible with the hydroloom package.
}
\description{
converts a compatible dataset into a fabric s3 class
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))

hy(x)

hy(x, clean = TRUE)[1:10,]

attr(hy(x), "orig_names")

}
