% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy_classes.R
\name{hy_capabilities}
\alias{hy_capabilities}
\title{What operations are available for this network?}
\usage{
hy_capabilities(x, include_conversions = FALSE)
}
\arguments{
\item{x}{object to query.}

\item{include_conversions}{logical. If TRUE, also marks functions reachable
via a single representation conversion.}
}
\value{
named logical vector.
}
\description{
Returns a named logical vector indicating which hydroloom
functions can operate on the input object given its current class and
columns.

Functions marked TRUE are directly callable. Use
\code{include_conversions = TRUE} to see what becomes available after
a single representation conversion.
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))
hy_capabilities(hy(x))

}
