% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hy.R
\name{hy_reverse}
\alias{hy_reverse}
\title{Reverse \code{hy} to Original Names}
\usage{
hy_reverse(x)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}
}
\value{
returns x with attribute names converted to original names provided to \link{hy}
}
\description{
renames hy object to original names and removes hy object
attributes.
}
\examples{
x <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))
x <- hy(x)

hy_reverse(x)

}
