% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_names.R
\name{hydroloom_name_definitions}
\alias{hydroloom_name_definitions}
\title{Hydroloom Name Definitions}
\description{
A names character vector containing definitions of all
attributes used in the hydroloom package.
}
\examples{
cat(paste0(names(hydroloom_name_definitions), ", ",
           hydroloom_name_definitions), sep = "\n")
}
