% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align_names.R
\name{hydroloom_name_definitions}
\alias{hydroloom_name_definitions}
\title{Hydroloom Name Definitions}
\value{
named character vector with \code{hydroloom_names} class to support custom print method
}
\description{
A names character vector containing definitions of all
attributes used in the hydroloom package.
}
\examples{
hydroloom_name_definitions
}
