% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_ids.R
\name{make_fromids}
\alias{make_fromids}
\title{DEPRECATED Convert "to" index ids to "from" index ids}
\usage{
make_fromids(index_ids, return_list = FALSE, upmain = NULL)
}
\arguments{
\item{index_ids}{data.frame as returned by \link{make_index_ids}}

\item{return_list}{logical if TRUE, the returned list will include a
"froms_list" element containing all from ids in a list form.}

\item{upmain}{data.frame containing \code{id} and \code{upmain} columns. \code{upmain} should
be a logical value indicating if the id is the upmain connection from its
downstream neighbors.}
}
\value{
list containing a "froms" matrix, "lengths" vector,
and optionally "froms_list" elements.
}
\description{
make_fromids is deprecated and will be removed in a future release.
Use \link{make_index_ids} with \code{mode = "from"} instead.
}
