% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_ids.R
\name{make_index_ids}
\alias{make_index_ids}
\alias{make_index_ids.data.frame}
\alias{make_index_ids.hy}
\title{make index ids}
\usage{
make_index_ids(x)

\method{make_index_ids}{data.frame}(x)

\method{make_index_ids}{hy}(x)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}
}
\value{
data.frame containing \code{indid} and \code{toindid} columns suitable for use
in fast graph traversal. If x is non-dendritic (\code{indid}:\code{toindid} is 1:many),
an adjacency matrix transformation is necessary to get \code{indid} to correspond
to rows. Use \link{format_nonden_toid} for this case.
}
\description{
makes index ids for the provided hy object. These can be used
for graph traversal algorithms such that the row number and id are equal.
}
\examples{

x <- data.frame(id = c(1, 2, 3, 4, 5, 6, 7, 8, 9),
              toid = c(2, 3, 4, 5, 0, 7, 8, 9, 4))

make_index_ids(x)

x <- hy(sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom")))

x <- add_toids(x, return_dendritic = FALSE)

x <- make_index_ids(x)

names(x)
class(x$to)
class(x$lengths)
class(x$to_list)
is.list(x$to_list$toindid)

}
