% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_utils.R
\name{make_index_ids}
\alias{make_index_ids}
\title{make index ids}
\usage{
make_index_ids(x)
}
\arguments{
\item{x}{hy data.frame}
}
\description{
makes index ids for the provided hy object. These can be used
for graph traversal algorithms such that the row number and id are equal.
}
\examples{
x <- hy(sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom")))

x <- add_toids(x)

index_ids <- make_index_ids(x)
}
