% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_ids.R
\name{make_index_ids}
\alias{make_index_ids}
\alias{make_index_ids.data.frame}
\alias{make_index_ids.hy}
\title{make index ids}
\usage{
make_index_ids(x, format = FALSE, complete = FALSE)

\method{make_index_ids}{data.frame}(x, format = FALSE, complete = FALSE)

\method{make_index_ids}{hy}(x, format = FALSE, complete = FALSE)
}
\arguments{
\item{x}{hy data.frame}

\item{format}{logical if TRUE, return will be a list containing an adjacency
matrix and a lengths vector indicating the number of connections from each node.}

\item{complete}{logical if TRUE return will also include a data.frame with an
\code{indid} column and a \code{toindid} list column. Has no affect if \code{format} is FALSE.}
}
\value{
data.frame containing \code{indid} and \code{toindid} columns suitable for use
in fast graph traversal. If x is non-dendritic (\code{indid}:\code{toindid} is 1:many),
an adjacency matrix transformation is necessary to get \code{indid} to correspond
to rows. Set \code{matrix=TRUE} to perform the transformation automatically.
}
\description{
makes index ids for the provided hy object. These can be used
for graph traversal algorithms such that the row number and id are equal.
}
\examples{

x <- data.frame(id = c(1, 2, 3, 4, 5, 6, 7, 8, 9),
              toid = c(2, 3, 4, 5, 0, 7, 8, 9, 4))

make_index_ids(x)

x <- hy(sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom")))

x <- add_toids(x)

make_index_ids(x)

}
