% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_network.R
\name{sort_network}
\alias{sort_network}
\alias{sort_network.data.frame}
\title{Get Sorted Network}
\usage{
sort_network(x, split = FALSE, outlets = NULL)

\method{sort_network}{data.frame}(x, split = FALSE, outlets = NULL)
}
\arguments{
\item{x}{data.frame network compatible with \link{hydroloom_names}.}

\item{split}{logical if TRUE, the result will be split into
independent networks identified by the id of their outlet. The
outlet id of each independent network is added as a "terminalID"
attribute.}

\item{outlets}{same as id in x; if specified only the network
emanating from these outlets will be considered and returned.}
}
\value{
data.frame containing a topologically sorted version
of the requested network and optionally a terminal id.
}
\description{
given a network with an id and and toid, returns a sorted
and potentially split set of output.

Can also be used as a very fast implementation of upstream
with tributaries navigation. The full network from each
outlet is returned in sorted order.
}
\examples{
g <- sf::read_sf(system.file("extdata/new_hope.gpkg", package = "hydroloom"))
g <- add_toids(g)
head(g <- sort_network(g))
g$topo_sort <- 1:nrow(g)
plot(g['topo_sort'])

}
